+++ lib/jin_widget_helper.dart
@@ -23,12 +23,13 @@ export './src/widgets/stream_handler.dart';
 export './src/widgets/value_observer.dart';
 
 //utils
-export './src/utils/jin_extension.dart';
-export './src/utils/jin_color_utils.dart';
+export './src/utils/after_build_mixin.dart';
+export './src/utils/form_mixin.dart';
 export './src/utils/form_validator.dart';
-export './src/utils/jin_utils.dart';
+export './src/utils/jin_color_utils.dart';
+export './src/utils/jin_extension.dart';
 export './src/utils/jin_navigator.dart';
+export './src/utils/jin_utils.dart';
+export './src/utils/jin_widget.dart';
 export './src/utils/no_splash_factory.dart';
 export './src/utils/page_navigator.dart';
-export './src/utils/jin_widget.dart';
-export './src/utils/after_build_mixin.dart';