+++ lib/src/widgets/action_button.dart
@@ -4,7 +4,7 @@ import 'package:jin_widget_helper/src/widgets/spacing.dart';
 
 class ActionButton extends StatelessWidget {
   ///recieve a ValueNotifier to indicate wether to show the loading or child
-  final ValueNotifier<bool> isLoading;
+  final ValueNotifier<bool> loadingNotifier;
   final Function onPressed;
   final Function onLongPressed;
   final double elevation;
@@ -20,21 +20,21 @@ class ActionButton extends StatelessWidget {
   final Widget child;
   final Widget icon;
 
-  ///if [stretch] is `true`, Button will take all remaining horizontal space
-  final bool stretch;
+  ///if [fullWidth] is `true`, Button will take all remaining horizontal space
+  final bool fullWidth;
 
   ///Create a button with loading notifier
   ActionButton({
     @required this.onPressed,
     @required this.child,
-    this.isLoading,
+    this.loadingNotifier,
     this.loadingWidget,
     this.color,
     this.icon,
     this.loadingColor = Colors.white,
     this.margin = const EdgeInsets.symmetric(vertical: 16),
     this.padding = const EdgeInsets.symmetric(vertical: 8, horizontal: 16),
-    this.stretch = true,
+    this.fullWidth = true,
     this.shape = const StadiumBorder(),
     this.onLongPressed,
     this.elevation = 2.0,
@@ -43,11 +43,11 @@ class ActionButton extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      width: stretch ? double.infinity : null,
+      width: fullWidth ? double.infinity : null,
       color: Colors.transparent,
       margin: margin,
       child: ValueListenableBuilder<bool>(
-        valueListenable: isLoading ?? ValueNotifier(false),
+        valueListenable: loadingNotifier ?? ValueNotifier(false),
         builder: (context, loading, _) {
           return RaisedButton(
             onPressed: loading ? () {} : onPressed,
@@ -73,10 +73,7 @@ class ActionButton extends StatelessWidget {
                   Container(
                     width: 20,
                     height: 20,
-                    child: CircularProgressIndicator(
-                      valueColor: AlwaysStoppedAnimation<Color>(loadingColor),
-                      strokeWidth: 3.0,
-                    ),
+                    child: JinWidget.platformLoadingWidget(color: loadingColor),
                   ),
             ),
           );