+++ lib/src/widgets/stream_handler.dart
@@ -45,7 +45,7 @@ class StreamHandler<T> extends StatelessWidget {
             ),
           );
         } else {
-          return Center(child: loading ?? JinWidget.platformLoadingWidget);
+          return Center(child: loading ?? JinWidget.platformLoadingWidget());
         }
       },
     );